# naicha-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Naicha Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/me/profile": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "获取当前用户资料",
        "operationId": "get_profile_api_v1_me_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/addresses": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "获取地址簿",
        "operationId": "list_addresses_api_v1_me_addresses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserAddressSchema"
                  },
                  "type": "array",
                  "title": "Response List Addresses Api V1 Me Addresses Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "me"
        ],
        "summary": "新增地址",
        "operationId": "create_address_api_v1_me_addresses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAddressCreateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAddressSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/me/addresses/{address_id}": {
      "put": {
        "tags": [
          "me"
        ],
        "summary": "更新地址",
        "operationId": "update_address_api_v1_me_addresses__address_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "address_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Address Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAddressUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAddressSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "me"
        ],
        "summary": "删除地址",
        "operationId": "delete_address_api_v1_me_addresses__address_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "address_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Address Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/loyalty/transactions": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "获取积分明细",
        "description": "获取当前用户的积分明细",
        "operationId": "get_loyalty_transactions_api_v1_me_loyalty_transactions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 10,
              "title": "Limit"
            },
            "description": "每页数量"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoyaltyTransactionsResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/coupons": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "获取优惠券列表",
        "description": "获取当前用户的优惠券列表",
        "operationId": "get_coupons_api_v1_me_coupons_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(active|used|expired|void)$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "筛选状态",
              "title": "Status"
            },
            "description": "筛选状态"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CouponsResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/orders": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "获取我的订单列表",
        "operationId": "list_my_orders_api_v1_me_orders_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Limit"
            },
            "description": "每页数量"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "订单状态筛选，多个状态用逗号分隔。允许值: pending_payment, paid, in_production, ready_for_pickup, completed, cancelled, refund_pending, refunded",
              "title": "Status"
            },
            "description": "订单状态筛选，多个状态用逗号分隔。允许值: pending_payment, paid, in_production, ready_for_pickup, completed, cancelled, refund_pending, refunded"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderResponseSchema"
                  },
                  "title": "Response List My Orders Api V1 Me Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/stamps": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "获取集点进度",
        "operationId": "get_stamp_status_api_v1_me_stamps_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StampStatusSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/orders": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "创建订单",
        "operationId": "create_order_api_v1_orders_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CouponSchema": {
        "properties": {
          "coupon_id": {
            "type": "integer",
            "title": "Coupon Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "meta_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Json"
          },
          "issued_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Used At"
          },
          "used_in_order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Used In Order Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "coupon_id",
          "user_id",
          "type",
          "status",
          "meta_json",
          "issued_at",
          "used_at",
          "used_in_order_id",
          "created_at"
        ],
        "title": "CouponSchema",
        "description": "优惠券"
      },
...
      "CouponsResponseSchema": {
        "properties": {
          "coupons": {
            "items": {
              "$ref": "#/components/schemas/CouponSchema"
            },
            "type": "array",
            "title": "Coupons"
          },
          "stats": {
            "$ref": "#/components/schemas/CouponStatsSchema"
          }
        },
        "type": "object",
        "required": [
          "coupons",
          "stats"
        ],
        "title": "CouponsResponseSchema",
        "description": "优惠券列表响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoyaltyTransactionSchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "delta_points": {
            "type": "integer",
            "title": "Delta Points"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "order_id",
          "delta_points",
          "reason",
          "created_at"
        ],
        "title": "LoyaltyTransactionSchema",
        "description": "积分明细记录"
      },
      "LoyaltyTransactionsResponseSchema": {
        "properties": {
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/LoyaltyTransactionSchema"
            },
            "type": "array",
            "title": "Transactions"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "current_points": {
            "type": "integer",
            "title": "Current Points"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "transactions",
          "total_count",
          "current_points",
          "limit",
          "offset"
        ],
        "title": "LoyaltyTransactionsResponseSchema",
        "description": "积分明细列表响应"
      },
...
      "OrderResponseSchema": {
        "properties": {
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "order_number": {
            "type": "string",
            "title": "Order Number"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "order_type": {
            "type": "string",
            "title": "Order Type"
          },
          "total_price": {
            "type": "number",
            "title": "Total Price"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "is_scheduled": {
            "type": "boolean",
            "title": "Is Scheduled"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "reminder_sent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Sent At"
          },
          "eta_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Minutes",
            "description": "预计等待/送达分钟数"
          },
          "eta_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Text",
            "description": "预计时间文案"
          },
          "pickup_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickup Code",
            "description": "取餐码（支付成功后生成）"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemSchema"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "order_number",
          "status",
          "order_type",
          "total_price",
          "created_at",
          "is_scheduled",
          "scheduled_at",
          "reminder_sent_at",
          "items"
        ],
        "title": "OrderResponseSchema"
      },
...
      "StampStatusSchema": {
        "properties": {
          "total_completed_orders": {
            "type": "integer",
            "title": "Total Completed Orders"
          },
          "stamps_in_cycle": {
            "type": "integer",
            "title": "Stamps In Cycle"
          },
          "rewards_available": {
            "type": "integer",
            "title": "Rewards Available"
          },
          "cycle_size": {
            "type": "integer",
            "title": "Cycle Size",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "total_completed_orders",
          "stamps_in_cycle",
          "rewards_available"
        ],
        "title": "StampStatusSchema",
        "description": "集点状态"
      },
...
      "UserAddressCreateSchema": {
        "properties": {
          "contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "address_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          }
        },
        "type": "object",
        "title": "UserAddressCreateSchema"
      },
      "UserAddressSchema": {
        "properties": {
          "address_id": {
            "type": "integer",
            "title": "Address Id"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "address_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "address_id",
          "contact_name",
          "phone",
          "address_line",
          "lat",
          "lng",
          "is_default",
          "created_at",
          "updated_at"
        ],
        "title": "UserAddressSchema"
      },
      "UserAddressUpdateSchema": {
        "properties": {
          "contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "address_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          }
        },
        "type": "object",
        "title": "UserAddressUpdateSchema"
      },
...
      "UserProfileSchema": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "loyalty_points": {
            "type": "integer",
            "title": "Loyalty Points"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "nickname",
          "avatar_url",
          "loyalty_points"
        ],
        "title": "UserProfileSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}